+++ rtl/bplib/arty/migui_arty_gsim.vhd
@@ -1,4 +1,4 @@
--- $Id: migui_arty_gsim.vhd 1181 2019-07-08 17:00:50Z mueller $
+-- $Id: migui_arty_gsim.vhd 1201 2019-08-10 16:51:22Z mueller $
 -- SPDX-License-Identifier: GPL-3.0-or-later
 -- Copyright 2018- by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 -- 
@@ -76,13 +76,22 @@ architecture sim of migui_arty is
   
 begin
 
+  -- On Arty we have
+  --   SYS_CLK_I   166.6 Mhz
+  --   controller  333.3 MHz
+  --   UI_CLK       83.3 MHz   (4:1)
+  -- therefore for simulation
+  --   f_vco    1000 MHz
+  --   --> mul   6        (f_vco/SYS_CLK)
+  --   --> div  12        (f_vco/UI_CLK)
+
   MIG_SIM : migui_core_gsim
     generic map (
       BAWIDTH => mig_bawidth,
       MAWIDTH => mig_mawidth,
       SAWIDTH => 24,
-      CLKMUI_MUL =>  7,
-      CLKMUI_DIV => 14)
+      CLKMUI_MUL =>  6,
+      CLKMUI_DIV => 12)
     port map (
       SYS_CLK             => SYS_CLK_I,
       SYS_RST             => SYS_RST,